implemented review feedback - moved functions mostly
